requirements.txt
python-telegram-bot==22.0
deep-translator==1.11.4
Flask==3.1.0
requests==2.32.5